Rewrite heptatonic scale generator
Previously, the generator would take a predetermined list of all of the
possible "white keys", transpose them starting on a given note, map the
given tones to those keys, and then ensure each note is spelled
correctly according to the starting note given.

Unfortunately the generator wasn't perfect and would end up generating
"invalid" sequences of notes whose spelling could not be determined.

This commit doesn't completely fix this but it does rewrite the
algorithm in a more sensible way. Instead of determining how notes
should be spelled on the fly, hardcode with all of the possible notes
and their spellings, transpose *this* list based on a given note and
then map the generated tones of the scale to this list.